Add files via upload
Table of values for allocation of excess catch in the general "rockfish genus" category during Covid to individual species.  The table pertains to species being assessed in 2025.  Values are for PC in 2020 and 2021 and PR in 2020.  Original data from RecFIN are shown for reference along with values to be allocated.  Original and allocation values should be added together.  See the README file dated 20230602.